requirements.txt
# Dependencias principales para el sistema de detección de estrés hídrico
earthengine-api>=0.1.400
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Manipulación y análisis de datos
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
xarray>=2023.1.0

# Aprendizaje automático
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# Procesamiento geoespacial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
folium>=0.14.0

# Solicitudes de API y manejo de datos
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.8.0

# Análisis de series temporales
statsmodels>=0.14.0
seasonal>=0.3.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuración y utilidades
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Validación de datos
cerberus>=1.3.0